refactor: reorganize message imports and add utility functions for color handling
